requirements.txt
cycler==0.10.0
kiwisolver==1.2.0
matplotlib==3.1.1
numpy==1.18.1
pandas==0.25.3
Pillow==7.2.0
pyparsing==2.4.7
gym==0.17.1
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
yellowbrick==1.1
joblib==0.16.0
seaborn==0.9.0
certifi==2020.6.20
mdptoolbox_hiive==4.0.3.1
mlrose==1.3.0
mlrose-hiive==2.1.3
dataframe-image==0.1.
dill==0.3.1.1